requirements.txt
# PDF Chat – requirements
# Install with: pip install -r requirements.txt

streamlit>=1.35.0
pypdf>=4.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.24.0
requests>=2.31.0
